+++ macros/src/function_like/println.rs
@@ -3,7 +3,7 @@ use proc_macro::TokenStream;
 use proc_macro2::TokenStream as TokenStream2;
 use proc_macro_error2::abort;
 use quote::quote;
-use syn::parse_macro_input;
+use syn::{parse_macro_input, parse_quote};
 
 use crate::construct;
 use crate::function_like::log::{Args, Codegen};
@@ -34,7 +34,8 @@ pub(crate) fn expand_parsed(args: Args) -> TokenStream2 {
         args.format_string.span(),
     );
 
-    let header = construct::interned_string(&format_string, "println", true, None);
+    let header =
+        construct::interned_string(&format_string, "println", true, None, &parse_quote!(defmt));
     let content = if exprs.is_empty() {
         quote!(
             defmt::export::acquire_header_and_release(&#header);